test+chore: finish the warning sweep and prune editor includes
An aborted build round hid a last warning set. The count was 46, not
zero. All are fixed now: vector constructors take size_t, pointer
arithmetic replaces subscript decay, unused locals and one unused
helper are deleted, file helpers become static, and the deliberate
exact-compare asserts carry file-top pragmas.

The editor no longer includes seven headers it never used directly:
ModDestMap, ModMatrixHighlight, ParamHelp, SynthParamLabels,
NoteStepControl, SeqLengthStepper, ModSourceCatalog. Symbol-level
checks confirmed zero direct uses; patch.h stays (the generator-pill
table uses its enums). A full rebuild verified no other TU needed a
transitive include added.

The verified state: zero compiler warnings and zero errors in Source
and tests. Only the vendored firmware reference and the vendored
clouds tree warn, both outside the edit scope.

+++ tests/apvts_test.cpp
@@ -105,7 +105,7 @@ double detectPitchHz (const std::vector<float>& x, double fs,
         return 0.0;
 
     // Normalise + compute the autocorrelation, tracking the global max.
-    std::vector<double> acf (maxLag + 2, 0.0);
+    std::vector<double> acf (static_cast<size_t> (maxLag + 2), 0.0);
     double globalMax = 0.0;
     for (int lag = minLag; lag <= maxLag; ++lag)
     {
